+++ packages/start-plugin-core/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @tanstack/start-plugin-core
 
+## 1.171.34
+
+### Patch Changes
+
+- Updated dependencies [[`7e93431`](https://github.com/TanStack/router/commit/7e93431ae9ff58c91c3c5ca10ffcb8414c1d0b13)]:
+  - @tanstack/router-core@1.171.22
+  - @tanstack/router-generator@1.167.28
+  - @tanstack/router-plugin@1.168.30
+  - @tanstack/start-server-core@1.169.26
+
 ## 1.171.33
 
 ### Patch Changes